impl(lv2): 사용자와 일정 간 연관관계 매핑

+++ src/main/java/com/example/scheduler2/service/ScheduleService.java
@@ -1,14 +1,18 @@
 package com.example.scheduler2.service;
 
 import com.example.scheduler2.domain.Schedule;
+import com.example.scheduler2.domain.User;
 import com.example.scheduler2.dto.ScheduleRequestDto.CreateScheduleDto;
 import com.example.scheduler2.dto.ScheduleRequestDto.UpdateScheduleDto;
 import com.example.scheduler2.dto.ScheduleResponseDto.ScheduleDetailDto;
 import com.example.scheduler2.repository.ScheduleRepository;
+import com.example.scheduler2.repository.UserRepository;
 import com.example.scheduler2.util.OptionalUtils;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -16,10 +20,13 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ScheduleService {
 
+    private final UserRepository userRepository;
     private final ScheduleRepository scheduleRepository;
 
-    public ScheduleDetailDto createSchedule(CreateScheduleDto createDto) {
+    public ScheduleDetailDto createSchedule(Long userId, CreateScheduleDto createDto) {
+        User user = OptionalUtils.getOrThrowNotFound(userRepository.findById(userId));
         Schedule schedule = new Schedule(createDto);
+        schedule.setUser(user);
         scheduleRepository.save(schedule);
         return new ScheduleDetailDto(schedule);
     }
@@ -36,14 +43,22 @@ public class ScheduleService {
     }
 
     @Transactional
-    public void updateSchedule(Long scheduleId, UpdateScheduleDto updateDto) {
+    public void updateSchedule(Long userId, Long scheduleId, UpdateScheduleDto updateDto) {
         Schedule schedule = OptionalUtils.getOrThrowNotFound(scheduleRepository.findById(scheduleId));
+        checkScheduleAuthor(userId, schedule.getUser());
         schedule.update(updateDto.getTitle(), updateDto.getContents());
     }
 
     @Transactional
-    public void deleteSchedule(Long scheduleId) {
+    public void deleteSchedule(Long userId, Long scheduleId) {
         Schedule schedule = OptionalUtils.getOrThrowNotFound(scheduleRepository.findById(scheduleId));
+        checkScheduleAuthor(userId, schedule.getUser());
         scheduleRepository.delete(schedule);
+    }
+
+    private static void checkScheduleAuthor(Long userId, User author) {
+        if (author == null || !userId.equals(author.getId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
     }
 }
